+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changes
+- `Update<T>` is now covariant in `T`.
+- `ModelInit<T, I>` is now covariant in `T`.
 
 ## [0.8.0] - 2019-11-19
 ### Changed